create mouse utils
